# BUG: add exception to searchsorted incompatible inputs