@luxedo 2024-06-10
Initial PR failed because of regression tests from more than 10 years ago. There are 3 tests:

### The first regression test passes:
https://github.com/numpy/numpy/blob/5aaede6eef94b35b926937129a4bff785d020269/numpy/_core/tests/test_regression.py#L859-L862

### The second test passes too:
https://github.com/numpy/numpy/blob/5aaede6eef94b35b926937129a4bff785d020269/numpy/_core/tests/test_regression.py#L2017-L2020
The new verification does not catches the error and the expected exception is thrown in a `getattr` inside `_wrapit`:
```python
File ~/code/os/numpy/numpy/_core/fromnumeric.py:46, in _wrapit(obj, method, *args, **kwds)
     43 # As this already tried the method, subok is maybe quite reasonable here
     44 # but this follows what was done before. TODO: revisit this.
     45 arr, = conv.as_arrays(subok=False)
---> 46 result = getattr(arr, method)(*args, **kwds)
     48 return conv.wrap(result, to_scalar=False)

TypeError: '<' not supported between instances of 'tuple' and 'float'
```

### The third test fails ([logs](https://github.com/numpy/numpy/actions/runs/9439431789/job/25997548313?pr=26650#step:9:772)) and is related to https://github.com/numpy/numpy/issues/642 and https://github.com/numpy/numpy/issues/2658.

https://github.com/numpy/numpy/blob/5aaede6eef94b35b926937129a4bff785d020269/numpy/_core/tests/test_regression.py#L2069-L2078

Which now throws:
```python
ValueError: Incompatible types for searching: a ((numpy.record, [('f0', '<i4'), ('f1', '<i4')])) and v (int64)
```

&nbsp;
## Summary
There's already a rudimentary type verification by means of not being able to compare types, but is not enough to catch `string` and `integer` discrepancies.

With this PR there will be two types of errors: 
1. The old `TypeError` by type comparison: `TypeError: '<' not supported between instances of 'tuple' and 'float'`
2. The new `ValueError` given `resolve_dtypes`: `TypeError: Incompatible types for searching: a (str) and v (int64)`

My suggestion from https://github.com/numpy/numpy/pull/26650/commits/647bb8583120a2513f57e454c86bb5a329fcf0c6 is to just use `TypeError` for the new exception as well, which I think makes sense. But I'm a bit bothered by the fact that there are two exceptions for handling incompatible data types.