@eendebakpt 2024-06-12
This PR has a negative impact on the performance for small arrays. A quick test:
```
import numpy as np
x=np.array([1, 2, 3])
y=np.array([1, 5])
z=np.searchsorted(x,y)

%timeit np.searchsorted(x,y)
``` 
Results in 
```
main: 753 ns ± 37.4 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
PR: 1.23 µs ± 34.2 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```
(that does not mean we should not merge this PR, as it does fix a bug)

* Could we add the type validation to `ndarray.searchsorted`? That might make the overhead a lot less 
* Are there any general guidelines on how much type checking we want to do?