@luxedo 2024-06-12
I agree!

Is there any opportunity for "private" functions and methods that already trust the argument types? If so, the work has to be done only once to bring the Python built-in types to numpy types and all the rest of the work can be done under the hood.