@luxedo 2024-06-18
Any updates on this? A very similar PR has been approved: https://github.com/numpy/numpy/pull/26667
Indeed there's a performance drop of around 60%, maybe placing the type validation in C code would make this faster.