@mattip 2024-07-02
Personally, I prefer the "first make it right, then make it fast" approach and am inclined to accept this PR as is. But if someone wants to dive into moving the `resolve_dtypes` call into C, the place to do it is in `PyArray_SearchSorted` https://github.com/numpy/numpy/blob/95073d138b4d3f2912f01f1442d203cf1844f0e1/numpy/_core/src/multiarray/item_selection.c#L2084

The implementation would still look up `np.less.resolve_dtypes` using the lookup-and-cache mechanism, and then call it with the already-converted arguments.