@eendebakpt 2024-07-03
@mattip The problem addressed in the PR is a bit on the boundary between fixing an bug, and providing a better error message, so I would prefer to solve it right in the first go. I would be willing to look in to the C implementation, but that will not be in the comings weeks, so if we do not want to wait there are no objections from my side to merge this.