@luxedo 2024-07-04
I can give a go at the C code as well but I'm very new to numpy's codebase, so a little guidance would be of much help.