@eendebakpt 2024-08-13
Performance considerations for this PR might have changed a bit since https://github.com/numpy/numpy/pull/27119. I am having trouble compiling on my own system, so I cannot provide any good benchmarks though.

@luxedo Do you still want to have a go at the C code? I think (but it would be nice if a numpy developer could confirm) that you could move the check to `PyArray_SearchSorted` in https://github.com/numpy/numpy/blob/main/numpy/_core/src/multiarray/item_selection.c#L2084. The input arguments have already been converted to arrays, and the dtypes are available via `PyArray_DESCR`