@luxedo 2024-09-02
Hi! I'm starting taking a look and trying to understand numpy's C code better. So far I got this:

```c
NPY_NO_EXPORT PyObject *
PyArray_SearchSorted(PyArrayObject *op1, PyObject *op2,
                     NPY_SEARCHSIDE side, PyObject *perm)
{
    /* Check if types are comparable */
   
    // Load 'less' ufunc to pass to py_resolve_dtypes_generic. DON'T KNOW HOW TO FIND THE ARGUMENTS.
    PyUFuncObject *ufunc_less = PyUFunc_FromFuncAndData("Don't know how to get 'less' ufunc");
    
    // Pack input types an a tuple to pass to py_resolve_dtypes_generic
    PyObject *dtypes = PyTuple_Pack(2, PyArray_DESCR(op1), op2);

    // Empty tuple as kwnames
    PyObject *kwnames = PyTuple_New(0);
   
     // Check if types are compatible
    PyObject *resolve = py_resolve_dtypes_generic(ufunc_less, NPY_FALSE, &dtypes, 2, kwnames);
    
    // Cleanup and early return if types are not comparable
    bool incompatible_types = resolve == NULL;
    Py_DECREF(dtypes);
    Py_DECREF(kwnames);
    Py_XDECREF(resolve);
    if (incompatible_types) {
        PyErr_SetString(PyExc_TypeError, "incomparable types");
        return NULL;
    }

  ...
```

This won't compile yet. Because I don't know how to get the `less` `ufunc` with `PyUFunc_FromFuncAndData` or any other way.

Is this in the correct path? Is there any other way to validate that the types can be compared?