@eendebakpt 2026-05-24
> What is the status on this PR?

The issue still exists. I will open a PR with an approach to resolve this in C (for performance reasons). If that turns out to be too complex, then we should merge this PR and accept the performance loss for now.